backend/requirements.txt
fastapi
uvicorn
requests
python-dotenv
google-genai
yfinance
portia-sdk-python[google]